% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cement.R
\name{cement}
\alias{cement}
\title{Hald's Cement Data}
\format{
A data.frame with 13 observations on the following 5 variables.
\describe{
  \item{x1}{percentage weight in clinkers of 3CaO.Al2O3}
  \item{x2}{percentage weight in clinkers of 3CaO.SiO2}
  \item{x3}{percentage weight in clinkers of 4CaO.Al2O3.Fe2O3}
  \item{x4}{percentage weight in clinkers of 2CaO.SiO2}
  \item{y}{heat evolved (calories/gram)}
}
}
\source{
Woods, H., Steinour, H. H. and Starke, H. R. (1932) Effect of composition of Portland cement on heat evolved during hardening. Industrial Engineering and Chemistry, 24, 1207–1214.
}
\description{
Heat evolved (cals/gm) in the setting of 13 samples of Portland cement with different percentage weight of chemical components.
}
\examples{
data("cement")
cement
}
