% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positiveStein.R, R/preliminaryTest.R
\name{coefficients.positiveStein}
\alias{coefficients.positiveStein}
\alias{coef.positiveStein}
\alias{coefficients.preliminaryTest}
\alias{coef.preliminaryTest}
\title{Extract Model Coefficients}
\usage{
\method{coefficients}{positiveStein}(object, ...)

\method{coef}{positiveStein}(object, ...)

\method{coefficients}{preliminaryTest}(object, ...)

\method{coef}{preliminaryTest}(object, ...)
}
\arguments{
\item{object}{An object of class "\code{positiveStein}", "\code{preliminaryTest}",
"\code{restricted}", "\code{Stein}" or "\code{unrestricted}".}

\item{...}{Other.}
}
\description{
coef is a generic function which extracts model
coefficients from objects returned by modeling functions.coefficients is an alias for it.
}
\examples{
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nr = 3, nc = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- positiveStein(X, y, H, h)
coefficients(model)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- positiveStein(X, y, H, h)
coef(model)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- preliminaryTest(X, y, H, h, alpha = 0.05)
coefficients(model)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- preliminaryTest(X, y, H, h, alpha = 0.05)
coef(model)
}
\seealso{
\code{\link{coefficients.positiveStein}}, \code{\link{coefficients.preliminaryTest}},
\code{\link{coefficients.restricted}}, \code{\link{coefficients.Stein}},
\code{\link{coefficients.unrestricted}}
\code{\link{coef.positiveStein}}, \code{\link{coef.preliminaryTest}},
\code{\link{coef.restricted}}, \code{\link{coef.Stein}},
\code{\link{coef.unrestricted}}.
}
