% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st.R
\name{predict.st}
\alias{predict.st}
\title{Extract Model Predictions Values}
\usage{
\method{predict}{st}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class "\code{stein}".}

\item{newdata}{An optional data frame in which to look for variables with which to predict.
If omitted, the fitted values are used.}

\item{...}{Other arguments.}
}
\value{
A vector of predictions.
}
\description{
Predicted values based on object \code{stein}.
}
\examples{
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nr = 3, nc = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- st(X, y, H, h)
predict(model, X)

}
\seealso{
\code{\link{predict.unres}},
\code{\link{predict.res}},
\code{\link{predict.pt}},
\code{\link{predict.ipt}},
\code{\link{predict.pst}}.
}
