% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unrestricted.R
\name{predict.unrestricted}
\alias{predict.unrestricted}
\title{Model Predictions}
\usage{
\method{predict}{unrestricted}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{unrestricted}.}

\item{newdata}{An optional data frame in which to look for variables with which to predict.
If omitted, the fitted values are used.}

\item{...}{Other arguments.}
}
\description{
\code{predict} is a generic function for predictions from the results of various
model fitting functions.
}
\examples{
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
model <- unrestricted(X, y)
predict(model, X)
}
\seealso{
\code{\link{predict.restricted}},
\code{\link{predict.preliminaryTest}},
\code{\link{predict.improvedpreliminaryTest}},
\code{\link{predict.stein}},
\code{\link{predict.positivestein}}.
}
