% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminaryTest.R
\name{preliminaryTest}
\alias{preliminaryTest}
\title{The preliminary test estimator}
\usage{
preliminaryTest(X, y, H, h, alpha)
}
\arguments{
\item{alpha}{float.}

\item{x}{scaler.}

\item{q}{scaler.}

\item{n}{number.}
}
\description{
positiveStein
}
\examples{
n <- 100
p <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simulate_data(n, p, beta)
X <- simulated_data$X
y <- simulated_data$y
# H beta = h
H <- matrix(c(1,1,-1,0,0,1,0,1,0,-1,0,0,0,1,0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, q)
preliminaryTest(X, y, H, h, alpha = 0.05)

# H beta != h
H <- matrix(c(1,1,-1,0,0,1,0,1,0,-1,0,0,0,1,0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(1, q)
preliminaryTest(X, y, H, h, alpha = 0.05)

}
\references{
ref
}
