% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restricted.R
\name{residuals.restricted}
\alias{residuals.restricted}
\title{residuals method for Model Fits}
\usage{
\method{residuals}{restricted}(object, ...)
}
\arguments{
\item{object}{An object of class "\code{restricted}".}

\item{...}{Other.}
}
\description{
residuals values based on model object.
}
\examples{
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- restricted(X, y, H, h)
residuals(model)
}
\seealso{
\code{\link{residuals.unrestricted}},
\code{\link{residuals.preliminaryTest}},
\code{\link{residuals.improvedpreliminaryTest}},
 \code{\link{residuals.stein}},
 #' \code{\link{residuals.positivestein}}.
}
