% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unres.R
\name{residuals.unres}
\alias{residuals.unres}
\title{Extract Model Residuals}
\usage{
\method{residuals}{unres}(object, ...)
}
\arguments{
\item{object}{An object of class \code{unrestricted}.}

\item{...}{Other arguments.}
}
\description{
Residuals values based on model object \code{unrestrcited}.
}
\examples{
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
model <- unres(X, y)
residuals(model)
}
\seealso{
\code{\link{residuals.res}},
\code{\link{residuals.pt}},
\code{\link{residuals.ipt}}
\code{\link{residuals.st}},
\code{\link{residuals.pst}}.
}
