% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restricted.R
\name{restricted}
\alias{restricted}
\title{The restricted estimator}
\usage{
restricted(X, y, H, h)
}
\arguments{
\item{x}{scaler.}

\item{q}{scaler.}

\item{n}{number.}

\item{alpha}{float.}
}
\description{
positiveStein
}
\examples{
n <- 100
p <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simulate_data(n, p, beta)
X <- simulated_data$X
y <- simulated_data$y
# H beta = h
H <- matrix(c(1,1,-1,0,0,1,0,1,0,-1,0,0,0,1,0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, q)
restricted(X, y, H, h)

# H beta != h
H <- matrix(c(1,1,-1,0,0,1,0,1,0,-1,0,0,0,1,0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(1, q)
restricted(X, y, H, h)

}
\references{
ref
}
