% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulation data}
\usage{
simulate_data(n, p, beta)
}
\arguments{
\item{n}{number.}

\item{x}{scaler.}

\item{q}{scaler.}

\item{alpha}{float.}
}
\description{
positiveStein
}
\examples{
n <- 100
p <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simulate_data(n, p, beta)
X <- simulated_data$X
y <- simulated_data$y

}
\references{
ref
}
