% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unrestricted.R
\name{unrestricted}
\alias{unrestricted}
\title{The Unrestricted estimator}
\usage{
unrestricted(X, y)
}
\arguments{
\item{x}{scaler.}

\item{q}{scaler.}

\item{n}{number.}

\item{alpha}{float.}
}
\description{
positiveStein
}
\examples{
n <- 100
p <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simulate_data(n, p, beta)
X <- simulated_data$X
y <- simulated_data$y
unrestricted(X,y)

}
\references{
ref
}
