% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImprovedpreliminaryTest.R
\name{coefficients.improvedpreliminaryTest}
\alias{coefficients.improvedpreliminaryTest}
\alias{coef.improvedpreliminaryTest}
\title{Extract Model Coefficients}
\usage{
\method{coefficients}{improvedpreliminaryTest}(object, ...)

\method{coef}{improvedpreliminaryTest}(object, ...)
}
\arguments{
\item{object}{An object of class "\code{improvedpreliminaryTest}".}

\item{...}{Other.}
}
\description{
\code{coef} is a generic function which extracts model
coefficients from objects returned by modeling \code{functions.coefficients}
is an alias for it.
}
\examples{
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- improvedpreliminaryTest(X, y, H, h, alpha = 0.05)
coefficients(model)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- improvedpreliminaryTest(X, y, H, h, alpha = 0.05)
coef(model)
}
\seealso{
\code{\link{coefficients.positivestein}}, \code{\link{coefficients.preliminaryTest}},
\code{\link{coefficients.restricted}}, \code{\link{coefficients.stein}},
\code{\link{coefficients.unrestricted}}, \code{\link{coefficients.improvedpreliminaryTest}},
\code{\link{coef.positivestein}}, \code{\link{coef.preliminaryTest}},
\code{\link{coef.restricted}}, \code{\link{coef.stein}},
\code{\link{coef.unrestricted}}, \code{\link{coef.improvedpreliminaryTest}}.
}
