% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positiveStein.R
\name{fitted.positiveStein}
\alias{fitted.positiveStein}
\title{Predict method for Linear Model Fits}
\usage{
\method{fitted}{positiveStein}(object, newdata)
}
\arguments{
\item{object}{An object of class "\code{positiveStein}", "\code{preliminaryTest}",
"\code{restricted}", "\code{Stein}" or "\code{unrestricted}".}

\item{newdata}{An optional data frame in which to look for variables with which to predict.
If omitted, the fitted values are used.}
}
\description{
Predicted values based on linear model object.
}
\examples{
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1,1,-1,0,0,1,0,1,0,-1,0,0,0,1,0), nr = 3, nc = p, byrow = TRUE)
h <- rep(0, nrow(H))
model<-positiveStein(X, y, H, h)
fitted(model)
}
\seealso{
\code{\link{predict.Arima}}, \code{\link{predict.bats}},
\code{\link{predict.tbats}}, \code{\link{predict.ets}},
\code{\link{predict.nnetar}}
\code{\link{fitted.Arima}}, \code{\link{fitted.bats}},
\code{\link{fitted.tbats}}, \code{\link{fitted.ets}},
\code{\link{fitted.nnetar}}.
}
