% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positiveStein.R, R/preliminaryTest.R
\name{fitted.positiveStein}
\alias{fitted.positiveStein}
\alias{predict.positiveStein}
\alias{fitted.preliminaryTest}
\alias{predict.preliminaryTest}
\title{Predict method for Model Fits}
\usage{
\method{fitted}{positiveStein}(object, newdata, ...)

\method{predict}{positiveStein}(object, newdata, ...)

\method{fitted}{preliminaryTest}(object, newdata, ...)

\method{predict}{preliminaryTest}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class "\code{positiveStein}", "\code{preliminaryTest}",
"\code{restricted}", "\code{Stein}" or "\code{unrestricted}".}

\item{newdata}{An optional data frame in which to look for variables with which to predict.
If omitted, the fitted values are used.}

\item{...}{Other.}
}
\description{
Predicted values based on model object.
}
\examples{
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nr = 3, nc = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- positiveStein(X, y, H, h)
fitted(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nr = 3, nc = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- positiveStein(X, y, H, h)
predict(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- preliminaryTest(X, y, H, h, alpha = 0.05)
fitted(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- preliminaryTest(X, y, H, h, alpha = 0.05)
predict(model, X)
}
\seealso{
\code{\link{predict.positiveStein}}, \code{\link{predict.preliminaryTest}},
\code{\link{predict.restricted}}, \code{\link{predict.Stein}},
\code{\link{predict.unrestricted}}
\code{\link{fitted.positiveStein}}, \code{\link{fitted.preliminaryTest}},
\code{\link{fitted.restricted}}, \code{\link{fitted.Stein}},
\code{\link{fitted.unrestricted}}.
}
