% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positivestein.R, R/preliminarytest.R,
%   R/restricted.R, R/stein.R, R/unrestricted.R
\name{fitted.positivestein}
\alias{fitted.positivestein}
\alias{predict.positivestein}
\alias{fitted.preliminaryTest}
\alias{predict.preliminaryTest}
\alias{fitted.restricted}
\alias{predict.restricted}
\alias{fitted.stein}
\alias{predict.stein}
\alias{fitted.unrestricted}
\alias{predict.unrestricted}
\title{Predict method for Model Fits}
\usage{
\method{fitted}{positivestein}(object, newdata, ...)

\method{predict}{positivestein}(object, newdata, ...)

\method{fitted}{preliminaryTest}(object, newdata, ...)

\method{predict}{preliminaryTest}(object, newdata, ...)

\method{fitted}{restricted}(object, newdata, ...)

\method{predict}{restricted}(object, newdata, ...)

\method{fitted}{stein}(object, newdata, ...)

\method{predict}{stein}(object, newdata, ...)

\method{fitted}{unrestricted}(object, newdata, ...)

\method{predict}{unrestricted}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class "\code{stein}, "\code{preliminaryTest}",
"\code{restricted}", "\code{positivestein}"" or "\code{unrestricted}".}

\item{newdata}{An optional data frame in which to look for variables with which to predict.
If omitted, the fitted values are used.}

\item{...}{Other.}
}
\description{
Predicted values based on model object.
}
\examples{
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- positivestein(X, y, H, h)
fitted(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- positivestein(X, y, H, h)
predict(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- preliminaryTest(X, y, H, h, alpha = 0.05)
fitted(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- preliminaryTest(X, y, H, h, alpha = 0.05)
predict(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- restricted(X, y, H, h)
fitted(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- restricted(X, y, H, h)
predict(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nr = 3, nc = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- stein(X, y, H, h)
fitted(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nr = 3, nc = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- stein(X, y, H, h)
predict(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
model <- unrestricted(X, y)
fitted(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
model <- unrestricted(X, y)
predict(model, X)
}
\seealso{
\code{\link{predict.positivestein}}, \code{\link{predict.preliminaryTest}},
\code{\link{predict.restricted}}, \code{\link{predict.stein}},
\code{\link{predict.unrestricted}}
\code{\link{fitted.positivestein}}, \code{\link{fitted.preliminaryTest}},
\code{\link{fitted.restricted}}, \code{\link{fitted.stein}},
\code{\link{fitted.unrestricted}}.
}
