% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImprovedpreliminaryTest.R, R/Stein.R,
%   R/positiveStein.R, R/preliminaryTest.R, R/restricted.R, R/unrestricted.R
\name{predict.improvedpreliminaryTest}
\alias{predict.improvedpreliminaryTest}
\alias{predict.stein}
\alias{predict.positivestein}
\alias{predict.preliminaryTest}
\alias{predict.restricted}
\alias{predict.unrestricted}
\title{Model Predictions}
\usage{
\method{predict}{improvedpreliminaryTest}(object, newdata, ...)

\method{predict}{stein}(object, newdata, ...)

\method{predict}{positivestein}(object, newdata, ...)

\method{predict}{preliminaryTest}(object, newdata, ...)

\method{predict}{restricted}(object, newdata, ...)

\method{predict}{unrestricted}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class "\code{unrestricted}", "\code{restricted}",
 "\code{preliminaryTest}", "\code{improvedpreliminaryTest}",
"\code{stein}" and"\code{positivestein}".}

\item{newdata}{An optional data frame in which to look for variables with which to predict.
If omitted, the fitted values are used.}

\item{...}{Other.}
}
\description{
\code{predict} is a generic function for predictions from the results of various
model fitting functions.
}
\examples{
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- improvedpreliminaryTest(X, y, H, h, alpha = 0.05)
predict(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nr = 3, nc = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- stein(X, y, H, h)
predict(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- positivestein(X, y, H, h)
predict(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- preliminaryTest(X, y, H, h, alpha = 0.05)
predict(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
p <- ncol(X)
# H beta = h
H <- matrix(c(1, 1, -1, 0, 0, 1, 0, 1, 0, -1, 0, 0, 0, 1, 0), nrow = 3, ncol = p, byrow = TRUE)
h <- rep(0, nrow(H))
model <- restricted(X, y, H, h)
predict(model, X)
n_obs <- 100
p_vars <- 5
beta <- c(2, 1, 3, 0, 5)
simulated_data <- simdata(n = n_obs, p = p_vars, beta)
X <- simulated_data$X
y <- simulated_data$y
model <- unrestricted(X, y)
predict(model, X)
}
\seealso{
\code{\link{predict.unrestricted}}, \code{\link{predict.restricted}},
 \code{\link{predict.preliminaryTest}}, \code{\link{predict.improvedpreliminaryTest}},
 \code{\link{predict.stein}}, \code{\link{predict.positivestein}}.
}
